reorganised components into dedicated folders
